moved all unit test to a new dir, and updated the path for jest to search for these unit test file
